+++ Examples/Multiple Carousels/iCarouselExampleViewController.m
@@ -110,8 +110,9 @@
     //create new view if no view is available for recycling
     if (view == nil)
     {
-        view = [[[UIImageView alloc] initWithImage:[UIImage imageNamed:@"page.png"]] autorelease];
-        view.layer.doubleSided = NO; //prevent back side of view from showing
+        view = [[[UIImageView alloc] initWithFrame:CGRectMake(0, 0, 200.0f, 200.0f)] autorelease];
+        ((UIImageView *)view).image = [UIImage imageNamed:@"page.png"];
+        view.contentMode = UIViewContentModeCenter;
         label = [[[UILabel alloc] initWithFrame:view.bounds] autorelease];
         label.backgroundColor = [UIColor clearColor];
         label.textAlignment = UITextAlignmentCenter;
@@ -140,6 +141,26 @@
     }
     
     return view;
+}
+
+- (CGFloat)carousel:(iCarousel *)carousel valueForOption:(iCarouselOption)option withDefault:(CGFloat)value
+{
+    //customize carousel display
+    switch (option)
+    {
+        case iCarouselOptionSpacing:
+        {
+            if (carousel == carousel2)
+            {
+                //add a bit of spacing between the item views
+                return value * 1.05f;
+            }
+        }
+        default:
+        {
+            return value;
+        }
+    }
 }
 
 @end