+++ Examples/Offsets Demo/iCarouselExampleViewController.m
@@ -122,7 +122,10 @@
 	//create new view if no view is available for recycling
 	if (view == nil)
 	{
-		view = [[[UIImageView alloc] initWithImage:[UIImage imageNamed:@"page.png"]] autorelease];
+		view = [[[UIImageView alloc] initWithFrame:CGRectMake(0, 0, 200.0f, 200.0f)] autorelease];
+        ((UIImageView *)view).image = [UIImage imageNamed:@"page.png"];
+        view.contentMode = UIViewContentModeCenter;
+        
 		label = [[[UILabel alloc] initWithFrame:view.bounds] autorelease];
 		label.backgroundColor = [UIColor clearColor];
 		label.textAlignment = UITextAlignmentCenter;
@@ -147,8 +150,8 @@
     {
         case iCarouselOptionSpacing:
         {
-            //reduce spacing slightly
-            return value * 0.9f;
+            //add a bit of spacing between the item views
+            return value * 1.05f;
         }
         default:
         {